Feito a regra e negócio para calcular a média de preço no MATERIAL

+++ src/main/java/br/com/siscompras/dto/MaterialDto.java
@@ -29,9 +29,9 @@ public class MaterialDto {
     private Long quantidade;
     private boolean status;
     private String andamento;
-    private BigDecimal mediavalor;
-    private BigDecimal minvalor;
-    private BigDecimal maxvalor;
+    private double mediavalor;
+    private double minvalor;
+    private double maxvalor;
     private String observacao;
     private String arquivo;
     private List<CotacaoDto> cotacaoDtos;
