+++ src/main/java/br/com/siscompras/dto/NovoMaterialDto.java
@@ -26,9 +26,9 @@ public class NovoMaterialDto {
     private String arquivo;
     private boolean status;
     private String andamento;
-    private BigDecimal mediavalor;
-    private BigDecimal minvalor;
-    private BigDecimal maxvalor;
+    private double mediavalor;
+    private double minvalor;
+    private double maxvalor;
 
 
     private UsuarioDto usuarioDto;
